% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSampler-method.R, R/step1fun.R
\name{GibbsSampler}
\alias{GibbsSampler}
\alias{GibbsSampler,}
\alias{GibbsSampler-method}
\alias{step1fun}
\alias{step1}
\title{Completes an iteration of the MCMC sampler in 4 steps.}
\usage{
GibbsSampler(df, J)

step1fun(y, n, J_items)
}
\arguments{
\item{df}{A data frame where each row represents a unique demographic profile.
Let \code{J} represent the total number of response items. Columns must be organized
 as follows:
\describe{
\item{\code{df[, ncol(df)]}}{The last column must be a vector containing the total number of
individuals in demographic profile i who responsed to item j.}
\item{\code{df[ , (ncol(df)-J):(ncol(df)-1)]}}{Columns containing counts
of affirmative responses per demographic profile/response item must appear between
those containing demographic characteristics and the column of totals described above.}
}}

\item{J}{A scalar representing the total number of response items.}

\item{y}{A matrix y with J columns corresponding to affirmative responses to each item,}

\item{n}{A vector of length N containing total respondents per unique demographic profile.}

\item{J_items}{A scalar representing the total number of response items.}
}
\value{
Describe final output mathematically
 \item{\code{final_output_placeholder}}{Practical description}

\eqn{\omega}: An NxJ matrix with elements \eqn{\omega_{ij} \sim PG(n_{ij},\mu_{ij}^{(t-1)})}
corresponding to demographic profiles i and response items j.
 \item{w}{The matrix \eqn{\omega} of Polya-Gamma-sampled values}
}
\description{
For internal use only. Runs the first step of the Gibbs sampler, which samples
\eqn{\omega_{ij} \sim PG(n_{ij},\mu_{ij}^{(t-1)})} for each unique demographic
profile/response item combination.
}
\note{
Need steps 4 and 5, need to replace for loops, need to verify parameters

Nested for loops need to be replaced for efficiency
}
\examples{




}
\seealso{
step1fun, step2fun, step3fun

step2fun, step3fun
}
\author{
Jacob Montgomery, Bryant Moy, Noa Dasanaike, Santiago Olivella

Jacob Montgomery, Bryant Moy, Noa Dasanaike, Santiago Olivella
}
\keyword{internal}
