% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2fun.R
\name{step2fun}
\alias{step2fun}
\alias{step2}
\title{step2fun}
\usage{
step2fun(X, omega, kappa, lambda)
}
\arguments{
\item{X}{An nx2 matrix with each column x_i = [\theta_i^{t-1},-1]}

\item{omega}{An nxn matrix defined as $diag({\omega_{ij}^{(t)}}_{i=1}^N)}

\item{kappa}{An nx1 matrix defined as the transpose of the jth row of the \Kappa NxN covariance matrix}

\item{lambda}{A 2x2 matrix defined as ???}
}
\value{
A 2x1 matrix \hat{\beta_j}:
 \item{\beta_j}{The sampled \beta_j from the distribution calculated from the above information}
}
\description{
Runs the second step of the Gibbs sampler, which samples the beta distribution created through the defined m and v once, given the input values
}
\note{
Currently still need to verify what exactly lambda is in this, add checks, and run a test
}
\examples{


}
\seealso{

}
\author{
Jacob Montgomery, Bryant Moy, Noa Dasanaike, Santiago Olivella
}
