% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4fun.R
\name{step4fun}
\alias{step4fun}
\alias{step4}
\title{step4fun}
\usage{
step4fun(K_rho, theta, sigma2_theta = 1)
}
\arguments{
\item{K_rho}{An N x N covariance matrix generated using a kernel of demographic features. NOTE: We need to either generate that here, or create a function for that}

\item{\eqn{\theta^{(t)}}}{An N x 2 matrix, which is the output of the previous step for all i.}
}
\value{

}
\description{
Runs the fourth step of the gibbs sampler, in which you sample \eqn{f^{(t)}}.
}
\note{
I think Alma's question about the output on theta is correct, and will affect the input section here.
}
\examples{


}
\seealso{
step1fun, step2fun, step3fun, GibbsSampler-method
}
\author{
Jacob Montgomery, Bryant Moy, Noa Dasanaike, Santiago Olivella
}
\keyword{internal}
