% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3fun.R
\name{step3fun}
\alias{step3fun}
\alias{step3}
\title{step3fun}
\usage{
step3fun(sigma, beta, f_prior, y)
}
\arguments{
\item{sigma}{A number, ?}

\item{beta}{A 2x1 matrix, the estimate of beta sampled in step 2}

\item{f_prior}{A number, ?}

\item{y}{A 2x1 matrix, ?}
}
\value{
A number \theta_i^{(t)}:
 \item{\theta_i^{(t)}}{The sampled \theta from the distribution}
}
\description{
Runs the third step of the Gibbs sampler, which samples \theta ideal point distribution given the funciton's inputs. The returned value is a number.
}
\note{
Need some clarity on the other values
}
\examples{


}
\seealso{

}
\author{
Jacob Montgomery, Bryant Moy, Noa Dasanaike, Santiago Olivella
}
